DataReader class
Implemented DataReader to read dictionary.txt
