+++ ConsoleApplication1/ConsoleApplication1.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30320.27
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ConsoleApplication1", "ConsoleApplication1.vcxproj", "{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}.Debug|x64.ActiveCfg = Debug|x64
+		{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}.Debug|x64.Build.0 = Debug|x64
+		{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}.Debug|x86.ActiveCfg = Debug|Win32
+		{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}.Debug|x86.Build.0 = Debug|Win32
+		{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}.Release|x64.ActiveCfg = Release|x64
+		{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}.Release|x64.Build.0 = Release|x64
+		{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}.Release|x86.ActiveCfg = Release|Win32
+		{8C831CB7-AEFF-4861-B53D-3587CEA8E0EE}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {826BD377-C850-4B3B-8836-A1A2D3904CE6}
+	EndGlobalSection
+EndGlobal
