+++ ConsoleApplication1/check_it.txt
@@ -1,0 +1,7 @@
+ This is a comlete transcript of Patel's decision, which was
+ issued orally, in which she explains the basis for granting
+ the injunction. Because it marks one of the first times a
+ judge has issued a deciasion in a case of online
+ music-swapping, Patel's interpreation of the lwa is likely
+ to be widely cited in subsequent cases and will be crucial
+ to the Court of Appeals as it grapples with the case.
