+++ ConsoleApplication1/check_it2.txt
@@ -1,0 +1,2 @@
+I am goin to write somethin here 😀
+	 Plese corrct all the wron words!
