+++ ConsoleApplication1/result For Check_2.txt
@@ -1,0 +1,2 @@
+This is a suggestion after spell checks 
+I(ai, hi, ic, id, ie, if, in, is, it, iv, pi, si, ti, a) am goin(groin, gin, coin, gain, goon, gown, grin, join, loin) to write somethin(something) here 😀(Cannot Find the suggestion) Plese(please, lese, plebe) corrct(correct) all the wron(wrong, won, worn, iron, pron, wren) words!
