+++ ConsoleApplication1/result For Check_it.txt
@@ -1,0 +1,2 @@
+This is a suggestion after spell checks 
+This is a comlete(complete, compete) transcript of Patel's(pates, pastel, plates, panels, patens, paters) decision, which was issued orally, in which she explains the basis for granting the injunction. Because it marks one of the first times a judge has issued a deciasion(decision) in a case of online(Cannot Find the suggestion) music-swapping,(Cannot Find the suggestion) Patel's(pates, pastel, plates, panels, patens, paters) interpreation(Cannot Find the suggestion) of the lwa(la, wa, law, lea) is likely to be widely cited in subsequent cases and will be crucial to the Court of Appeals as it grapples with the case.
