--- SpellChecker.hpp
@@ -1,18 +1,0 @@
-//Kyeore Heo 
-//91kyoheo@gmail.com
-#ifndef SpellChecker_hpp
-#define SpellChecker_hpp
-#include <string>
-#include <vector>
-
-class SpellChecker
-{
-public:
-    SpellChecker();
-    SpellChecker(const std::string& inputDictionary);
-    ~SpellChecker();     
-private:
-
-};
-
-#endif /* SpellCheck_hpp */
