--- check_it2.txt
@@ -1,2 +1,0 @@
-I am goin to write somethin here 😀
-	 Plese corrct all the wron words!
